backend/requirements.txt
# FastAPI и сервер
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.0

# База данных
psycopg[binary]==3.1.19

# JWT - пароли
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Утилиты
python-dotenv==1.0.1

# AI модуль 
numpy==1.26.4
scikit-learn==1.5.2

# Redis 
redis==5.2.1